+++ src/main/java/ru/cft/StringType.java
@@ -3,7 +3,7 @@ package ru.cft;
 import lombok.Getter;
 
 public enum StringType {
-    INTEGER("-?[0-9]+"), DOUBLE("-?[0-9]+(.|,)[0-9]+"), STRING(".*");
+    INTEGER("-?[0-9]+"), FLOAT("-?[0-9]+(.|,)[0-9]+"), STRING(".*");
 
     @Getter
     private final String regexp;
